Changed install.sh, higher sleep tor circuit
Tor is moved up in the install.sh script. This means it now has a
shorter time before it is used for data transfers.
To make sure there is a circuit before the install continues; the
sleep for a Tor circuit has been raised from
120 seconds to 300 seconds. It is a very conservative/high number.
But it is an unattended install, so this is acceptable for now.

Might look into a more efficient way to do this.