tlsdate-script.sh use short cmd options
tlsdate has issues with the long command line version of the options:
'--proxy' and '--host'. The '--proxy' option will fail with the following
errors:
"SSL connection failed"
"child process failed in SSL handshake".
The '--host' option will not give an error. But in the --verbose option you
can see that no hostname is set/used.

So for now the short versions for '--proxy' and '--host' will be used.
Which are '-x' and '-H'.